--- a/.gitignore
+++ b/.gitignore
@@ -236,6 +236,7 @@ lcov*.info
 /rb_mjit_header.h
 /mjit_config.h
 /include/ruby-*/*/rb_mjit_min_header-*.h
+/mjit_instruction.rb
 
 # /wasm/
 /wasm/tests/*.wasm