--- a/inits.c
+++ b/inits.c
@@ -104,6 +104,8 @@ rb_call_builtin_inits(void)
     BUILTIN(marshal);
 #if USE_MJIT
     BUILTIN(mjit);
+    BUILTIN(mjit_compiler);
+    BUILTIN(mjit_instruction);
 #endif
     Init_builtin_prelude();
 }