--- a/ruby.c
+++ b/ruby.c
@@ -1563,10 +1563,22 @@ ruby_opt_init(ruby_cmdline_options_t *opt)
 
     rb_warning_category_update(opt->warn.mask, opt->warn.set);
 
+#if USE_MJIT
+    // rb_call_builtin_inits depends on RubyVM::MJIT.enabled?
+    if (opt->mjit.on)
+        mjit_enabled = true;
+#endif
+
     Init_ext(); /* load statically linked extensions before rubygems */
     Init_extra_exts();
     rb_call_builtin_inits();
     ruby_init_prelude();
+
+#if USE_MJIT
+    // mjit_init is safe only after rb_call_builtin_inits defines RubyVM::MJIT::Compiler
+    if (opt->mjit.on)
+        mjit_init(&opt->mjit);
+#endif
 
     ruby_set_script_name(opt->script_name);
     require_libraries(&opt->req_list);
@@ -1914,12 +1926,6 @@ process_options(int argc, char **argv, ruby_cmdline_options_t *opt)
 
     ruby_gc_set_params();
     ruby_init_loadpath();
-
-#if USE_MJIT
-    if (opt->mjit.on)
-        /* Using TMP_RUBY_PREFIX created by ruby_init_loadpath(). */
-        mjit_init(&opt->mjit);
-#endif
 
     Init_enc();
     lenc = rb_locale_encoding();