--- a/tool/mjit/.gitignore
+++ b/tool/mjit/.gitignore
@@ -1,0 +1,1 @@
+/Gemfile.lock