--- a/tool/update-deps
+++ b/tool/update-deps
@@ -120,7 +120,7 @@ FILES_NEED_VPATH = %w[
   known_errors.inc
   lex.c
   miniprelude.c
-  mjit_compile.inc
+  mjit_compile_attr.inc
   newline.c
   node_name.inc
   opt_sc.inc