… lib/mjit/c_64.rb
… lib/mjit/c_pointer.rb
… lib/mjit/compiler.rb
… mjit_compiler.c
… mjit_compiler.rb
… template/Makefile.in
… tool/mjit/bindgen.rb
… tool/ruby_vm/views/_mjit_compile_getconstant_path.erb
… tool/ruby_vm/views/_mjit_compile_insn.erb
… tool/ruby_vm/views/_mjit_compile_insn_body.erb
… tool/ruby_vm/views/_mjit_compile_invokebuiltin.erb
… tool/ruby_vm/views/_mjit_compile_ivar.erb
… tool/ruby_vm/views/_mjit_compile_pc_and_sp.erb
… tool/ruby_vm/views/_mjit_compile_send.erb
… tool/ruby_vm/views/mjit_compile.inc.erb
… tool/ruby_vm/views/mjit_instruction.rb.erb
… win32/Makefile.sub