Rewrite the MJIT compiler in Ruby for maintenance reasons.

Because we're running the compiler logic in a forked child process, the performance impact seems not too significant:

<details>
<summary>Benchmark results</summary>

```
before: ruby 3.2.0dev (2022-09-05T02:43:14Z master 45fe7f7575) +MJIT [x86_64-linux]
after: ruby 3.2.0dev (2022-09-05T02:56:23Z ruby-mjit ceee54cffc) +MJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
30k_ifelse     2864.7       9.8         2611.3      0.8         1.10          1.08
30k_methods    7841.3       6.2         6814.5      4.7         1.15          1.04
activerecord   106.7        0.3         104.5       0.2         1.02          1.02
binarytrees    148.5        2.2         148.8       2.8         1.00          1.12
cfunc_itself   41.3         0.3         35.9        0.3         1.15          1.05
chunky_png     488.1        27.5        497.7       27.0        0.98          1.01
erubi          208.0        23.7        215.8       25.1        0.96          0.99
erubi_rails    33.9         11.5        30.6        22.6        1.11          0.66
fannkuchredux  3615.0       4.3         3588.0      0.0         1.01          1.01
fib            39.3         2.0         37.7        0.1         1.04          1.25
getivar        94.9         13.4        76.7        0.1         1.24          1.90
keyword_args   117.7        0.4         125.3       1.4         0.94          1.08
lee            777.4        27.4        815.5       27.3        0.95          0.99
liquid-render  170.2        13.4        177.0       12.2        0.96          0.83
mail           156.2        13.2        159.3       13.1        0.98          0.96
nbody          43.7         19.3        41.8        19.2        1.04          1.13
optcarrot      2934.6       4.2         2892.8      3.7         1.01          0.95
psych-load     1412.0       28.4        1439.0      31.5        0.98          0.95
railsbench     1589.8       0.9         1655.0      2.6         0.96          0.99
respond_to     119.1        0.3         132.7       0.6         0.90          1.03
rubykon        4631.0       17.5        5324.3      19.7        0.87          0.88
setivar        54.6         7.8         60.7        2.7         0.90          1.42
str_concat     32.6         36.3        31.5        33.3        1.04          1.93
-------------  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```

</details>

The C binding for MJIT is generated by libclang and [ffi-clang](https://github.com/ioquatix/ffi-clang). libclang is already a source development dependency for YJIT, so nothing new. ffi-clang is a new external dependency, but I trust the maintainer, Samuel. I'm gonna set up CI to maintain this like YJIT in a separate PR. Once it's ready, you should be able to just copy-paste its output to fix MJIT in the same way as YJIT.